requirements.txt
openai==1.3.0
pandas==2.0.0
python-dotenv==1.0.0
openpyxl==3.10.0
requests==2.31.0
pytest==7.4.0
